Logging added. Game updated to 2.2

+++ src/main/kotlin/model/ship/components/Engine.kt
@@ -1,14 +1,14 @@
 package model.ship.components
 
-import model.ship.Requirements
 import kotlinx.serialization.Serializable
+import model.ship.Requirements
 
 @Serializable
 data class Engine(
     val symbol: String,
     val name: String,
     val description: String,
-    val condition: Long? = null,
+    val condition: Float? = null,
     val speed: Long,
     val requirements: Requirements
 )
