+++ src/main/kotlin/model/ship/components/Reactor.kt
@@ -1,14 +1,14 @@
 package model.ship.components
 
-import model.ship.Requirements
 import kotlinx.serialization.Serializable
+import model.ship.Requirements
 
 @Serializable
 data class Reactor(
     val symbol: String,
     val name: String,
     val description: String,
-    val condition: Long? = null,
+    val condition: Float? = null,
     val powerOutput: Long,
     val requirements: Requirements
 )
